pyproject.toml
[tool.poetry]
name = "site-crawler"
version = "0.1.0"
description = "Crawl a site and check for things"
authors = ["mshafer1 <2565361+mshafer1@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.17.2"
selenium-requests = "^2.0.3"
beautifulsoup4 = "^4.12.3"
python-decouple = "^3.8"
click = "^8.1.7"
black = "^24.1.1"
tqdm = "^4.67.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-snapshot = "^0.9.0"
ni-python-styleguide = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
